Step 9.19: Extract DatePicker component from inline App.jsx definition
