+++ src/App.jsx
@@ -89,6 +89,7 @@ import useKeyboardShortcuts from './hooks/useKeyboardShortcuts.js';
 import { DayPlannerContext } from './context/DayPlannerContext.jsx';
 import FrameNudgeCard from './components/FrameNudgeCard.jsx';
 import DeadlinePickerPopover from './components/DeadlinePickerPopover.jsx';
+import DatePicker from './components/DatePicker.jsx';
 import DesktopLayout from './components/DesktopLayout.jsx';
 import MobileLayout from './components/MobileLayout.jsx';
 import ShortcutHelpModal from './components/ShortcutHelpModal.jsx';
@@ -3999,132 +4000,6 @@ const DayPlanner = () => {
   };
 
   // DeadlinePickerPopover extracted to src/components/DeadlinePickerPopover.jsx
-
-
-  // FIX 2: Add missing DatePicker component
-  const DatePicker = ({ value, onChange, onClose }) => {
-    const [viewDate, setViewDate] = useState(() => {
-      if (value) {
-        const parts = value.split('-');
-        return new Date(parseInt(parts[0]), parseInt(parts[1]) - 1, 1);
-      }
-      return new Date();
-    });
-
-    const getDaysInMonth = () => {
-      const year = viewDate.getFullYear();
-      const month = viewDate.getMonth();
-      const firstDay = new Date(year, month, 1);
-      const lastDay = new Date(year, month + 1, 0);
-      const daysInMonth = lastDay.getDate();
-      const startingDayOfWeek = firstDay.getDay();
-      
-      const days = [];
-      // Add empty slots for days before the first of the month
-      for (let i = 0; i < startingDayOfWeek; i++) {
-        days.push(null);
-      }
-      // Add all days in the month
-      for (let i = 1; i <= daysInMonth; i++) {
-        days.push(new Date(year, month, i));
-      }
-      return days;
-    };
-
-    const changeMonth = (delta) => {
-      setViewDate(new Date(viewDate.getFullYear(), viewDate.getMonth() + delta, 1));
-    };
-
-    const monthNames = ['January', 'February', 'March', 'April', 'May', 'June', 
-                       'July', 'August', 'September', 'October', 'November', 'December'];
-    const days = getDaysInMonth();
-    const today = dateToString(new Date());
-
-    return (
-      <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-[60]" onClick={onClose} onKeyDown={(e) => { if (e.key === 'Escape') { e.preventDefault(); e.stopPropagation(); onClose(); } }} tabIndex={-1} ref={(el) => el && el.focus()}>
-        <div
-          className={`${cardBg} rounded-lg shadow-xl p-6 ${borderClass} border max-w-sm w-full mx-4`}
-          onClick={(e) => e.stopPropagation()}
-        >
-          <div className="flex items-center justify-between mb-4">
-            <button
-              onClick={() => changeMonth(-1)}
-              className={`p-2 rounded ${hoverBg}`}
-            >
-              <ChevronLeft size={20} className={textSecondary} />
-            </button>
-            <h3 className={`text-lg font-semibold ${textPrimary}`}>
-              {monthNames[viewDate.getMonth()]} {viewDate.getFullYear()}
-            </h3>
-            <button
-              onClick={() => changeMonth(1)}
-              className={`p-2 rounded ${hoverBg}`}
-            >
-              <ChevronRight size={20} className={textSecondary} />
-            </button>
-          </div>
-
-          <div className="grid grid-cols-7 gap-1 mb-2">
-            {['Su', 'Mo', 'Tu', 'We', 'Th', 'Fr', 'Sa'].map(day => (
-              <div key={day} className={`text-center text-sm font-semibold p-2 ${textSecondary}`}>
-                {day}
-              </div>
-            ))}
-          </div>
-
-          <div className="grid grid-cols-7 gap-1">
-            {days.map((day, index) => {
-              if (!day) {
-                return <div key={`empty-${index}`} className="p-2"></div>;
-              }
-              const dayStr = dateToString(day);
-              const isSelected = dayStr === value;
-              const isToday = dayStr === today;
-
-              return (
-                <button
-                  key={index}
-                  onClick={() => {
-                    onChange(dayStr);
-                    onClose();
-                  }}
-                  className={`p-2 text-center rounded-lg transition-colors ${
-                    isSelected 
-                      ? 'bg-blue-600 text-white font-bold' 
-                      : isToday
-                        ? darkMode ? 'bg-blue-900 text-blue-200 font-semibold' : 'bg-blue-100 text-blue-900 font-semibold'
-                        : darkMode 
-                          ? 'hover:bg-gray-700 text-gray-300' 
-                          : 'hover:bg-stone-100 text-stone-700'
-                  }`}
-                >
-                  {day.getDate()}
-                </button>
-              );
-            })}
-          </div>
-
-          <div className="flex gap-2 mt-4">
-            <button
-              onClick={() => {
-                onChange(dateToString(new Date()));
-                onClose();
-              }}
-              className="flex-1 px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700"
-            >
-              Today
-            </button>
-            <button
-              onClick={onClose}
-              className={`flex-1 px-4 py-2 rounded-lg ${darkMode ? 'bg-gray-700' : 'bg-stone-200'} ${textPrimary} ${hoverBg}`}
-            >
-              Cancel
-            </button>
-          </div>
-        </div>
-      </div>
-    );
-  };
 
   // Voice input — parse and add callbacks (must be after allTags is defined)
   // Build a text summary of existing tasks for AI context
